Added 40% Pause Menu Logic, will make the Settings GUI Soon
I still need to bind WASD to its respective directions so I can put it in the keybinds settings, I also need to add pause keybinds to that + mouse dpi and inverted.

I also need logic to save them to the player's computer so they dont need to set it again and again.